Conflicts:
	integtest/multi_output_file_test.py

+++ integtest/large_trigger_record_test.py
@@ -1,7 +1,14 @@
+# two of the goals of this test are to verify that...
+# * if the TR size is slightly over half of the maximum HDF5 file size, then
+#   we see one TR per file
+# * if the TR size is larger than the maximum HDF5 file size, we also see one
+#   TR per file.
+
 import pytest
 import os
 import re
 import urllib.request
+import copy
 
 import dfmodules.data_file_checks as data_file_checks
 import integrationtest.log_file_checks as log_file_checks
@@ -10,28 +17,35 @@ import dfmodules.integtest_file_gen as integtest_file_gen
 
 # Values that help determine the running conditions
 number_of_data_producers=10
-run_duration=23  # seconds
+run_duration=32  # seconds
 number_of_readout_apps=3
 number_of_dataflow_apps=1
-trigger_rate=0.2 # Hz
+trigger_rate=0.06 # Hz
 token_count=1
-readout_window_time_before=9000000
+readout_window_time_before=10000000
 readout_window_time_after=1000000
-data_rate_slowdown_factor=10
+data_rate_slowdown_factor=1
 
 # Default values for validation parameters
 expected_number_of_data_files=4
 check_for_logfile_errors=True
-expected_event_count=4
+expected_event_count=1
 expected_event_count_tolerance=1
-wib1_frag_hsi_trig_params={"fragment_type_description": "WIB",
-                           "hdf5_detector_group": "TPC", "hdf5_region_prefix": "APA",
-                           "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
-                           "min_size_bytes": 3712080, "max_size_bytes": 3712080}
+wibeth_frag_55pct_params={"fragment_type_description": "WIBEth",
+                             "fragment_type": "WIBEth",
+                             "hdf5_source_subsystem": "Detector_Readout",
+                             "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                             "min_size_bytes": 38678472, "max_size_bytes": 38678472}
+wibeth_frag_125pct_params={"fragment_type_description": "WIBEth",
+                               "fragment_type": "WIBEth",
+                               "hdf5_source_subsystem": "Detector_Readout",
+                               "expected_fragment_count": (number_of_data_producers*number_of_readout_apps),
+                               "min_size_bytes": 91411272, "max_size_bytes": 91411272}
 triggercandidate_frag_params={"fragment_type_description": "Trigger Candidate",
-                              "hdf5_detector_group": "Trigger", "hdf5_region_prefix": "Region",
+                              "fragment_type": "Trigger_Candidate",
+                              "hdf5_source_subsystem": "Trigger",
                               "expected_fragment_count": 1,
-                              "min_size_bytes": 72, "max_size_bytes": 216}
+                              "min_size_bytes": 72, "max_size_bytes": 280}
 hsi_frag_params ={"fragment_type_description": "HSI",
                              "fragment_type": "Hardware_Signal",
                              "hdf5_source_subsystem": "HW_Signals_Interface",
@@ -50,14 +64,10 @@ confgen_name="daqconf_multiru_gen"
 hardware_map_contents = integtest_file_gen.generate_hwmap_file(number_of_data_producers, number_of_readout_apps)
 
 conf_dict = config_file_gen.get_default_config_dict()
-try:
-  urllib.request.urlopen('http://localhost:5000').status
-  conf_dict["boot"]["use_connectivity_service"] = True
-except:
-  conf_dict["boot"]["use_connectivity_service"] = False
 conf_dict["readout"]["data_rate_slowdown_factor"] = data_rate_slowdown_factor
 conf_dict["readout"]["use_fake_data_producers"] = True
-conf_dict["readout"]["clock_speed_hz"] = 50000000
+conf_dict["readout"]["clock_speed_hz"] = 62500000 # DuneWIB/WIBEth
+conf_dict["readout"]["eth_mode"] = True # WIBEth
 conf_dict["trigger"]["trigger_rate_hz"] = trigger_rate
 conf_dict["trigger"]["trigger_window_before_ticks"] = readout_window_time_before
 conf_dict["trigger"]["trigger_window_after_ticks"] = readout_window_time_after
@@ -67,14 +77,19 @@ conf_dict["dataflow"]["apps"] = [] # Remove preconfigured dataflow0 app
 for df_app in range(number_of_dataflow_apps):
     dfapp_conf = {}
     dfapp_conf["app_name"] = f"dataflow{df_app}"
+    dfapp_conf["max_file_size"] = 2*1024*1024*1024
     conf_dict["dataflow"]["apps"].append(dfapp_conf)
 
-confgen_arguments={"Base_System": conf_dict,
+oversize_conf = copy.deepcopy(conf_dict)
+oversize_conf["trigger"]["trigger_window_before_ticks"] = readout_window_time_before * 2.5
+
+confgen_arguments={"TRSize_55PercentOfMaxFileSize": conf_dict,
+                   "TRSize_125PercentOfMaxFileSize": oversize_conf,
                   }
 # The commands to run in nanorc, as a list
 nanorc_command_list="integtest-partition boot conf".split()
-nanorc_command_list+="start_run 101 wait ".split() + [str(run_duration)] + "stop_run --wait 2 wait 2".split()
-nanorc_command_list+="start 102 wait 3 enable_triggers wait ".split() + [str(run_duration)] + "stop_run wait 2".split()
+nanorc_command_list+="start_run --wait 10 101 wait ".split() + [str(run_duration)] + "stop_run --wait 2 wait 2".split()
+nanorc_command_list+="start 102 wait 10 enable_triggers wait ".split() + [str(run_duration)] + "stop_run wait 2".split()
 nanorc_command_list+="scrap terminate".split()
 
 # The tests themselves
@@ -100,7 +115,10 @@ def test_data_files(run_nanorc):
     local_expected_event_count=expected_event_count
     local_event_count_tolerance=expected_event_count_tolerance
     fragment_check_list=[triggercandidate_frag_params, hsi_frag_params]
-    fragment_check_list.append(wib1_frag_hsi_trig_params)
+    if "trigger_window_before_ticks" in run_nanorc.confgen_config["trigger"].keys() and run_nanorc.confgen_config["trigger"]["trigger_window_before_ticks"] > 15000000:
+        fragment_check_list.append(wibeth_frag_125pct_params)
+    else:
+        fragment_check_list.append(wibeth_frag_55pct_params)
 
     # Run some tests on the output data file
     assert len(run_nanorc.data_files)==expected_number_of_data_files
@@ -115,9 +133,6 @@ def test_data_files(run_nanorc):
             assert data_file_checks.check_fragment_sizes(data_file, fragment_check_list[jdx])
 
 def test_cleanup(run_nanorc):
-    print("============================================")
-    print("Listing the hdf5 files before deleting them:")
-    print("============================================")
     pathlist_string=""
     filelist_string=""
     for data_file in run_nanorc.data_files:
@@ -125,13 +140,18 @@ def test_cleanup(run_nanorc):
         if str(data_file.parent) not in pathlist_string:
             pathlist_string += " " + str(data_file.parent)
 
-    os.system(f"df -h {pathlist_string}")
-    print("--------------------")
-    os.system(f"ls -alF {filelist_string}");
+    if pathlist_string and filelist_string:
+        print("============================================")
+        print("Listing the hdf5 files before deleting them:")
+        print("============================================")
 
-    for data_file in run_nanorc.data_files:
-        data_file.unlink()
+        os.system(f"df -h {pathlist_string}")
+        print("--------------------")
+        os.system(f"ls -alF {filelist_string}");
 
-    print("--------------------")
-    os.system(f"df -h {pathlist_string}")
-    print("============================================")
+        for data_file in run_nanorc.data_files:
+            data_file.unlink()
+
+        print("--------------------")
+        os.system(f"df -h {pathlist_string}")
+        print("============================================")
